backend/requirements.txt
# Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9

# ML
scikit-learn==1.6.1
xgboost==2.1.0
lightgbm==4.5.0
shap==0.45.0
imbalanced-learn==0.14.1
numpy==1.26.4
pandas==2.2.0
scipy==1.13.0

# LLM & Agents
anthropic==0.34.0

# Data
duckdb==1.1.0
pdfplumber==0.11.0

# Web Research
tavily-python==0.5.0
beautifulsoup4==4.12.3

# PDF Generation
weasyprint==62.3
jinja2==3.1.4
markdown==3.7

# Task Queue
redis==5.1.0

# WebSocket
websockets==13.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utils
httpx==0.27.0
aiofiles==24.1.0
python-dateutil==2.9.0
